\name{checkMeasErrorModel}
\alias{checkMeasErrorModel}
\title{Helper function for gbev-function.}
\description{
A help function for checking the specification of the measurement error model.
}
\usage{
checkMeasErrorModel(measErrorModel)
}
\arguments{
\item{measErrorModel}{A list of components specifying the measurement error model.}
}
\keyword{hplot}



