\name{treeMatrix}
\alias{treeMatrix}
\title{Matrix with boosted trees.}
\description{
A matrix containing all of the boosted trees is returned.
}
\usage{
treeMatrix(object,deci=3)
}
\arguments{
\item{object}{A fitted object of type \code{gbev.object}.}
\item{deci}{number of decimal points of reals.}
}
\value{
Returns a matrix with all the trees. 
}
\keyword{ hplot }
